output
Enter a sentence 
saveetha school of engineering
No. of vowels in saveetha school of engineering = 12

Enter a sentence 
india is my country
No. of vowels in india is my country = 6


Enter a sentence 
all are my brothers and sisters
No. of vowels in all are my brothers and sisters = 8


why dry sky
No. of vowels in whr dry sky = 0

shy try cry
No. of vowels in shy try cry = 0

education
No. of vowels in education = 5